Refactor: Move commitment tree types to dedicated module
Move CommitmentTreeData, CommitmentTreeRoots, and CommitmentTreeSizes
from the monolithic types.rs (3281 lines) to a dedicated commitment.rs
module. This is the first step in breaking down the large types file
into focused, manageable modules.

Changes:
- Create chain_index/types/commitment.rs with moved types
- Add module declaration and re-exports to types.rs
- Remove original type definitions from types.rs
- Maintain exact same functionality and API

Part of types modularization effort to improve code organization.
